udev: Implement a way to group devices of a same tree
This commit adds the `groupRecursive` option to the udev device
discovery handler, this configuration switch allows to group devices
that share a parent/child relation.

This can be used to group into one akri instance a complex device that
is split into multiple nodes. It is up to the user to define the
boundaries of the device as there are no standard way to do so.

It works by comparing sysfs path for a device prefixing another one
(i.e `/devices/pci0` is a parent of `/devices/pci0/0.1/0.1.1/0.1.1:1`,
this example is obviously fictive)

It adds all devnodes to the mounts and add a UDEV_DEVNODE_n property for
each one of them.

The UDEV_DEVPATH just need the topmost one as all other are
subdirectories of it.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>