udev: Add devnode to mounts not devpath
The udev discovery handler currently adds the udev devpath to
the device mounts, this doesn't allow access to the devnode and moreover
said devpath is not a correct absolute path (it is relative to `/sys`).

This commit changes the current behavior to add the devnode (if any) to
the device mounts instead of the (wrong) sysfs devpath.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>