controller: Retry when watch errors occurs
According to Kube-rs documentation, when a watch error occurs (e.g
because of the kube API being currently unavailable), the `Watcher`
still forwards the error, but will retry if the receiver continues
to wait for a new event.

This commit changes how the `Watcher` is used in order to handle those
errors and let the `Watcher` retry. A backoff mechanism is also setup so
we don't end up retrying forever.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>