Make agent less likely to restart on kube API error
This commit changes the way config are watched and the events handled,
this change have, as main purpose, to ensure the agent won't restart if
it looses the connection to the kube API. Two mechanisms are employed to
do this:
 - Change the Watcher use to make it re-try in case of error (same
   change that was done in 846662eae8af82d8546d03637bcf551084cb73b4 for
   the controller)
 - Handle the `Restarted` event so that in case of de-sync with API we
   are still able to recover without restarting (i.e handle
   creation/modification and deletion of `Configuration resources that
   occured during the connection loss).

The latter required some refactor of the `ConfigMap` to ensure we have
all the required information to do delete operation without having the
related event (after a `Restart` event only the current state of the
resources are available, not their events).

This refactor also fixed a yet undetected issue: We were unable to
handle two `Configuration` resources with the same name even accross
different namespaces.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>