udev: Use device specs instead of mounts
Device plugin mechanism diffrentiate Mounts, and DeviceSpecs, Mounts
should be used to make libraries or tools available to the pod, and
DeviceSpecs to make device nodes available to the pod.

This commit switch from using Mounts to using DeviceSpec for device
nodes. This currently makes Mounts unused.

This is related to #145

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>