controller: Stop panicking when unable to create pod.
Currently the controller panic if it encounters an error while trying to
create a pod/job for an instance, thus causing an impact far wider than
just the failing instance.

This should only have impact on the said instance, so as a first step,
this commit stop the controller from panicking in this case and just
output a log entry.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>