feat: Cleanup metadata
In multi-node setups, instances are not deleted correctly. The
Agent fails to submit the updated spec due to metadata fields being
present. Also, finalizers are not cleared properly, so Instances
remain stale even if the devices are not discovered.

We create a new instance, keeping only the spec, cleaning up the
nodes.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>